BIC; updated demo
